+++ frontend/lib/components/AddBrainModal/AddBrainModal.tsx
@@ -3,7 +3,7 @@ import { FormProvider, useForm } from "react-hook-form";
 import { defaultBrainConfig } from "@/lib/config/defaultBrainConfig";
 import { BrainConfig } from "@/lib/types/brainConfig";
 
-import { AddBrainConfig } from "./AddBrainConfig";
+import { AddBrainConfig } from "./components/AddBrainConfig/AddBrainConfig";
 
 type AddBrainModalProps = {
   triggerClassName?: string;
@@ -17,10 +17,8 @@ export const AddBrainModal = ({
   });
 
   return (
-    <>
-      <FormProvider {...methods}>
-        <AddBrainConfig triggerClassName={triggerClassName} />
-      </FormProvider>
-    </>
+    <FormProvider {...methods}>
+      <AddBrainConfig triggerClassName={triggerClassName} />
+    </FormProvider>
   );
 };
