+++ frontend/lib/components/AddBrainModal/components/AddBrainConfig/index.ts
@@ -1,0 +1,1 @@
+export * from "./AddBrainConfig";
