--- frontend/lib/components/AddBrainModal/styles.css
@@ -1,3 +1,0 @@
-.foo {
-    white-space: 'pre-wrap';
-}
