+++ frontend/lib/types/brainConfig.ts
@@ -2,9 +2,9 @@ export const brainStatuses = ["private", "public"] as const;
 
 export type BrainStatus = (typeof brainStatuses)[number];
 
-export const knowledgeSources = ["doc", "api"] as const;
+export const brainTypes = ["doc", "api"] as const;
 
-export type KnowledgeSource = (typeof knowledgeSources)[number];
+export type BrainType = (typeof brainTypes)[number];
 
 export type BrainConfig = {
   model: Model;
@@ -18,7 +18,7 @@ export type BrainConfig = {
   supabaseKey?: string;
   prompt_id?: string;
   status: BrainStatus;
-  knowledgeSource: KnowledgeSource;
+  brainType: BrainType;
   prompt: {
     title: string;
     content: string;
